pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kangni-agents"
version = "0.1.0"
description = "AI agent system"
authors = [{name = "duxin"}]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "mysql-connector-python>=9.4.0",
    "sqlalchemy>=2.0.0",
    "sqlalchemy-utils>=0.41.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "mcp>=1.13.1",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "sqlparse>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kangni_agents"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "flake8",
    "mypy",
]